Only search processes within terminal emulators
The process search is the bottleneck in terms of speed. Since vim is
likely to be running in the terminal, the search can be sped up by only
searching processes controlled by a terminal.

GUIs (e.g. gvim) should still work because the GUI program should be the
focused process, which is not restricted to only those running in the
terminal.